fixarr/requirements.txt
tmdbsimple
cinemagoer
pyyaml
requests
pytest
python-dotenv
rapidfuzz
pymediainfo
